v0.45.0: updated install/run
