+++ setup.py
@@ -107,6 +107,7 @@ setuptools.setup(name=NAME,
                  entry_points={
                      'console_scripts': [
                          'exotic = exotic.exotic:main',
+                         'exotic-gui = exotic-gui.exotic-gui:main',
                      ],
                  }
                  )
